test: move unsubscribe all tests to unsubscribe file
